depricated: customer and payload factory class
