+++ src/Contract/FactoryInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Flutterwave\Contract;
+
+interface FactoryInterface
+{
+    public function create(array $data): Entityinterface;
+
+}
