+++ src/Service/ApplePay.php
@@ -7,9 +7,10 @@ namespace Flutterwave\Service;
 use Flutterwave\Contract\ConfigInterface;
 use Flutterwave\Contract\Payment;
 use Flutterwave\EventHandlers\ApplePayEventHandler;
-use Flutterwave\Payload;
+use Flutterwave\Entities\Payload;
 use Flutterwave\Traits\Group\Charge;
 use GuzzleHttp\Exception\GuzzleException;
+use Psr\Http\Client\ClientExceptionInterface;
 use stdClass;
 
 class ApplePay extends Service implements Payment
@@ -39,9 +40,10 @@ class ApplePay extends Service implements Payment
     }
 
     /**
+     * @param Payload $payload
      * @return array
      *
-     * @throws GuzzleException
+     * @throws ClientExceptionInterface
      */
     public function charge(Payload $payload): array
     {
