+++ src/Service/CardPayment.php
@@ -4,13 +4,15 @@ declare(strict_types=1);
 
 namespace Flutterwave\Service;
 
+use Exception;
 use Flutterwave\Contract\ConfigInterface;
 use Flutterwave\Contract\Payment;
 use Flutterwave\EventHandlers\CardEventHandler;
-use Flutterwave\Payload;
+use Flutterwave\Entities\Payload;
 use Flutterwave\Traits\Group\Charge;
 use GuzzleHttp\Exception\GuzzleException;
 use InvalidArgumentException;
+use Psr\Http\Client\ClientExceptionInterface;
 
 class CardPayment extends Service implements Payment
 {
@@ -41,7 +43,7 @@ class CardPayment extends Service implements Payment
      * @return array
      * @throws GuzzleException
      */
-    public function initiate(\Flutterwave\Payload $payload): array
+    public function initiate(Payload $payload): array
     {
         if (self::$count >= 2) {
             //TODO: if payload does not have pin on 2nd request, trigger a warning.
@@ -64,11 +66,12 @@ class CardPayment extends Service implements Payment
     }
 
     /**
+     * @param Payload $payload
      * @return array
      *
-     * @throws GuzzleException
+     * @throws ClientExceptionInterface
      */
-    public function charge(\Flutterwave\Payload $payload): array
+    public function charge(Payload $payload): array
     {
         $tx_ref = $payload->get('tx_ref');
         $this->logger->notice("Card Service::Started Charging Card tx_ref:({$tx_ref})...");
@@ -118,7 +121,10 @@ class CardPayment extends Service implements Payment
     }
 
     /**
-     * @throws \Exception
+     * @param \stdClass $response
+     * @param $payload
+     * @return array
+     * @throws Exception
      */
     public function handleAuthState(\stdClass $response, $payload): array
     {
