+++ src/Service/Customer.php
@@ -5,24 +5,25 @@ declare(strict_types=1);
 namespace Flutterwave\Service;
 
 use Flutterwave\Contract\CustomerInterface;
-use Flutterwave\Customer as Person;
+use Flutterwave\Entities\Customer as Person;
+use Flutterwave\Factories\CustomerFactory;
 use InvalidArgumentException;
 
-class Customer implements CustomerInterface
+/**
+ * Class Customer.
+ * @deprecated use \Flutterwave\Factories\CustomerFactory instead
+ */
+class Customer
 {
+    protected CustomerInterface $customerFactory;
+
+    public function __construct()
+    {
+        $this->customerFactory = new CustomerFactory();
+    }
+
     public function create(array $data = []): Person
     {
-        $data = array_change_key_case($data);
-        if (empty($data)) {
-            throw new InvalidArgumentException('Customer data is empty');
-        }
-
-        $person = new Person();
-        $person->set('fullname', $data['full_name']);
-        $person->set('email', $data['email']);
-        $person->set('phone_number', $data['phone']);
-        $person->set('address', $data['address'] ?? null);
-
-        return $person;
+           return $this->customerFactory->create($data);
     }
 }
